requirements.txt
addict==2.2.0
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
coveralls==1.6.0
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
django-filters==0.2.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
isort==4.3.15
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
markup==0.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
